Anomaly_Detection using Python
1st Project
Anamoly_detection: It is a geo-spatial data and I have extracted the data from .gpkg to csv format and performed data exploratoty analysis and data visualization on the data to understand each dataset distribution and relationship to detect if there is any kind of anomaly.


2nd Project
Anomaly detection using Isolation Forest:
For the csv file loaded, we are trying to aggregate the data by load_Date and then based on the count fo rows, we are finding the anomaly. I am trying to wrap it in the function so that we can even try to create a module out of it and then run it for any data with the same format.